Update all Funcionalidades features

+++ src/main/java/org/example/estufaprodmanager/services/FuncionalidadesService.java
@@ -5,9 +5,9 @@ import org.example.estufaprodmanager.repositories.FuncionalidadesRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 @Service
 public class FuncionalidadesService {
@@ -21,17 +21,38 @@ public class FuncionalidadesService {
     }
 
     // List by UUID
-    public Optional<Funcionalidades> getById(String id){
+    public Optional<Funcionalidades> getById(UUID id){
         return repository.findById(id);
     }
 
     // Create
-    public Funcionalidades save(Funcionalidades entity){
-        return entity;
+    public Funcionalidades create(Funcionalidades funcionalidade){
+        return repository.save(funcionalidade);
+    }
+
+    // Update
+    public Funcionalidades update(UUID id, Funcionalidades novaFuncionalidade){
+        Optional<Funcionalidades> funcionalidadeExistente = repository.findById(id);
+
+        if(funcionalidadeExistente.isPresent()){
+
+            Funcionalidades funcionalidade = funcionalidadeExistente.get();
+
+            funcionalidade.setNome(novaFuncionalidade.getNome());
+            funcionalidade.setDescricao(novaFuncionalidade.getDescricao());
+
+            return repository.save(funcionalidade);
+        } else {
+            throw new RuntimeException("Funcionalidade não encontrada com o ID: " + id);
+        }
     }
 
     // Delete
-    public void delete(String id){
-        repository.deleteById(id);
+    public void delete(UUID id){
+        if (repository.existsById(id)){
+            repository.deleteById(id);
+        } else {
+            throw new RuntimeException("Erro ao apagar funcionalidade");
+        }
     }
 }
